data-validator/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.3.4
jsonschema==4.21.1
python-dotenv==1.0.0
pydantic==2.5.3
